requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0

# Configuration & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
